requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
flake8==3.8.4
future==0.18.2
GPUtil==1.4.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
oyaml==1.0
packaging==20.7
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-box==5.2.0
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
rope==0.18.0
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
snowballstemmer==2.0.0
toml==0.10.2
torch==1.7.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.1
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
